add a new disassembler
this diassembler uses a cache to store the movement possibilities for
each piece pair and relative position of this pair. The cache is then
later on used to faster fill the matrix.

This algorithm could be improved by keeping the cache over the period
of solving one puzzle. This is escpecially useful when the puzzle has
many assemblies (like Easy living)

This algorithm speeds up complex puzzles like Pelikan Burr quite a bit
(only 50% time left) but slows down puzzles with many simple assemblies
as the cache has to be rebuild each time.