implement piece groups
This allows you to put piece into groups that don't need to be
taken apart to solve the disassembly of the puzzle. This is required
for example for puzzles like "cube in cage" where the cage consists
of 3 movable pieces that can not be taken apart.

The gui currently alows you to increase the group number.

Group number 0 is special, it means that the piece doesn't belong to
any group at all and must be single in the disassembled puzzle