add hotspot functionality
a Hotspot is a point inside a voxel space that always stays at the
same position reltvie to the shape defined inside the voxel space,
so if you have the hotspot at some special corner and now flip rotate
and shift the piece around inside the voxelspace the hotspot will stay
at the same corner of the piece and move along with this corner

What is that useful for?

To rotate assemblies. because now it is possible to recalculate the
orientation of a piece and calculate the new position of the hotspot
independent of the concrete shape of the piece