improved the disassembler a bit
the disassembler now no longer finds moves where one piece is moved
x units and another piece is moved y units in the same direction

the disassembler now finds moved where 2 pieces are moved in the same
direction, until now it only found these when movement of the first
piece forced the 2nd to move.

This needs to be improved to larger groups of pieces.