speedup for reduction
the idea here is that if one piece occupies certain unit in all
its placements it is impossible that any other piece can
occupy that unit, so all placements for any other piece that
uses that voxel can be removed. This lets the matrix shink
must faster and saves us some of the costly cover and uncover
operations. The results can be quite enormous: One of frans
polycube puzzle where you put the pieces onto sticks went down
from 377 to 27 seconds.