fix bug with multi pieces that prevented finding solutions sometimes
This bug occured in the multiPieceWhy puzzle file in the regression
tests. THe problem was that the program expects to have at least
n possible placements for a piece that occures n times, otherwise
it is not possible to place them all. But the special first piece
might have only one placement removing that from all the others
and thus all other pieces might have only n-1 placement and that is
still valid ad the last placement is in the first piece