lower changes for the occurance of a fatal problem with the assembler
the problem occures when a multi-piece is selected as symmetry breaker
and later on a voxel is selected that can not be filled with the 1st
of the multi-piece and thus the 2nd gets selected.

The "solution" right now is to try to not choose a multi-piece as symmetry
breaker even if the symmetry conditions are better.

Now the problem can only occure in problems that have only multi pieces