fix in rotation check
this fixes a problem, that occures, when the pivot piece contains
mirror symmetries. Then the comparison works even, if there are
mirrored shapes inside the transformed assemblies. Of course mirrored
shapes are not available and so the assembly is invalid

There is, of course the possibility to swap to pieces that are the
mirror of one another. This needs to be inserted, but I will not
do that right here