virtualize transform
the transform function is really the work horse when it comes to generating alls
possible orientations. It it the only function used for that by the libraray. the rotxyz
functions are only used by the gui. So make transform virtual so that each class can
have its own implementations as spheres nees another aproach and can not work with
the current scheme of adding up several rotxyz calls. This also removes the rotxyz members
from sym, as nobody needs them the rot arrays are now part of the voxel_x_c classes

later on rotxyz will be removed from the voxel_c class as they are really dependend
on the grid space if they do make a sense at all