huge speedup for layouter
first when a layouter widget was resized it used to call the Fl_Group resize function, which first
placed all widgets resulting in recursive layouter call for all those placements and then the 
layouter placed the widgets again resulting in another recursive cascade. That is removed now by
only calling the Fl_Widget resize

second the minimum size of a layouter stays constant normally, so cache that value for later calls
to the function.