add the new dancing link assembler
this assembler is __WAY__ faster then the old one if it comes to multipieces
it is a bis slower when only single pieces are involved. That is why the old
one will live on. The old assembler will be simplified to only handle single
pieces once the new assembler is working properly.

So what works: it can assemble.

Placement browser is not working, the reduction is not yet up to paar with the
old assembler (the fast line removal is missing)

The assembler is seleced in a factory function in the gridType_c class. In case
it is neccessary to remove it just remove the selection of the new
assembler in this class and make the canHandle function of the old assembler
return always true

Right now there are also quite some commented out debug prints left. They will vanish
over time...