make BurrTools handle unplaced pieces
this has always been prepared but never used and so it never properly worked. A piece
is placed if its transformation is not 0xFF. This is now properly wrapped and checked
in all placed where necessary (at least I hope so). 

The patch also increases the possible number of pieces from 255 to MAXINT. This limit
was caused by using an unsigned char as index into the placement...