add a counting optimisation
the idea behind this is that if we have piece ranges and fixed pieces to place we do know
how many voxels can be contributed by range pieces. An example, 10 piece with 6 cubes plus
one 4 cube piece shall make a 4x4x4 cube. We want to try all different 4 cube pieces possible.
This is done by adding all 4 cube pieces with a 0-1 range.
Without the optimisation the programm will add several of the 4 cube pieces because there still
are more than 10 placements for the 6-unit pieces.
Now we add an additional column that contains the number of missing cubes as condition (in the
example 4) and a weight for each of the range pieces that is equal to the number of units that
piece is made of (again 4 in this example).
The result is that the program will not add more range pieces than are possible due to the number
of units the result has.