first step for voxelDrawer cleanup
this patch spearates the drawer from the organisation part. We do now have the drawer
which is different for each different space grid and only contains some drawing functions
this part is exchanged when a new space grid is selected instead of the whole widget

the 2nd part is the voxelFrame. This part organizes the shapes and calls the draw functions
it has a drawer object that is calls for each grid dependend operation