more memory efficient disassembler node + get rid of that >10000 rubbish for removed nodes
Now the disassembler can only work with puzzles up to 128 units in size because the positions
of pieces can not be bigger, but the memory savings are probably more important, even the largest
puzzles I know don't get that big

the >10000 was used to show when pieces are removed. This didn't work any longer with such a small
value range, so a better scheme (transformation + direction) is used. Thas big number scheme is
still valid for the disassembly, so we need to work on that later on