separate puzzle from problem
this is a first in a set of cleanup patches for the huge puzzle class
it divides the puzzle in a puzzle class that handles a vector of
problems. The problem was already used internally but never exposed

Now the problem itself is used wherever possible instead of the puzzle
problem index pair

This is a huge patch with more than 800 changes, I hope I didnt mess up
too much. I also didn't do the GUI update to nicely because there will be
a new GUI anyway, so no need to be to fuzzy about that

additionally the following things are done

- remove the color constraint special class and replace the STL set template class
- make some more functions inline
- lots of comments in the header file
- make the puzzle pointer a reference as there MUST be a puzzle
- remove the insanity of 0xFFFFFFs for num Assemblies Solutions and usedTime, it now works with the solverState
- the burrgrowing has been #if nulled, too much bit rot