fix a problem with the mirror filter
When the 2 mirror pieces do have a different range assigned it can happen
that after mirroring the now mirrored piece is outside of its valid range
and thus the new assembly is invalid. The old code still used that assembly
for the "is smaller" comparison trhowing away the valid base because of
the new invalid assembly. It expected that this invalid assembly would be
found but of course that is not the case. This resulted in missing solutions