Add config option for rapid 3D editing
Rapid 3D editing allows the user to wiggle the mouse when
adding/removing voxels to perform repeat actions very fast. It was
originally added in 8782ebd30cc99323257ff4fe5702b0da56d2986c. This is
useful for editing very large puzzles, but it's much too easy to do
accidentally.

This adds a config option for rapid 3D editing, which is off by default.