Undo / redo functionality
This implements undo and redo by saving the full state of the puzzle
after each change. Under the file menu is now an option to undo and
redo, with keyboard shortcuts Ctrl+Z and Ctrl+Y.

Undo history is saved in XML format for now, since that was easiest way
to make the implementation obviously correct. We may need to do
something more intelligent in the future, as this takes more time and
space than simply copying the data structure, even though that will be
more difficult.

In order to avoid thread safety issues, undo and redo is disabled while
the solve thread is running. Undo states are also not recorded while the
solve thread is running. This means that performing an undo after a
solve finishes will undo all changes performed during the solve. This
isn't usually a problem since most ways to edit the puzzle are disabled
while solving. See the comment in undomanager.h for a full discussion of
handling thread safety.

Along with the actual puzzle state, we also store a message about what
edit was made which is displayed later when the user performs an undo or
redo.